Follow the change of ExplainState.
ExplainState was changed at 9.4.1 and 9.3.6. This change breaks where
using es->grouping_stack so change the way to access it so that one
source supports both before and after the change was made. After all,
for instance, PG 9.4.0 and 9.4.1 cannot be able to share the same
pg_store_plans binary.