Make compile quiet phase 2
MacOS X Maverics and later offers only LLVM compiler (clang) and it
warns for implicit type conversions between enums. This program uses
some enum types for a member in the same struct type but they cannot
be merged because one of them is defined in PostgreSQL. The situation
is safe as long as an enumeration and a word table are in one-to-one
correspondence so I decided to change the type of the member to
integer. This change make the commit
78415c006cd24f41d722a41014c526b43731968f useless so it is reverted
along with this commit.

Reported by Fujii Masao.