Fix text plan representation for sort keys
Sort keys are in a list but read as a simple string. As the result
only the last of the list is shown in a text representation. This
commit fixes it.  This patch adds new setter macro LIST_SETTER for
this purpose and Output: is also changed to make use of the macro.