Change the way to treat nested json-objects
Previously the text converting parser assumes that all JSON object are
Plan or a children of Plans but Target Tables breaks the
assumption. This patch changes the way to detect the end of a Plan (or
a child of Plans) cleaner so that it can handle JSON objects other
than Plan.