Remove the queryid_stat_statements field.
With pg14 and above core postgres can compute a queryid without any external
extension, so we can simply rely on it.  We do so by asking for the core to
compute a queryid if users allow it.  Note that if users explicitly disable
compute_query_id, we will obey and not record anything.

For previous versions, automatically detect if a queryid is computed by some
external extensions or not.  If yes, use it as the key queryid otherwise
fallback on internal queryid calculation.

Bump PGSP_FILE_HEADER as this breaks the storage format.