Introduce file storage of plan texts
pg_stat_statements has been changed so that it separately stores query
texts into a file outside the main hash table. This configuration is
advantageous in several aspects. Let pg_store_plans make use the same
technique.  We don't intend to obsolete the existing in-hash-entry
plan text storage but remove that feature once so that this feature is
planted more easily.