Restore in-hash plan text storage as an alternative
We once moved completely to store plan texts into a file.  This commit
re-introduces the in-hash plan storage as an alternative choice with
the file storage, in case where garbage collection on files is
painful.

The in-hash storage is fixed-length so the efficiency of memory usage
if rather low.  We could improve the efficiency by using the same
strategy with the file storage if DSA, or expandable shared memory,
were available on postmaster, but it is not available at least for
now.  So as the second best we stay providing the same method with the
existing implement.